Update strategy-open-first.md
Line 30 - changed 'solutions' to plural. "...by building solutions together." Alternative could be "...by building THE solution together."
Line 46 - added quick line to lead in to free/open source software explanations.
Line 65 - Is it worth expanding on "locked in"? Maybe... 'locked in to a restricted way of thinking/working, processes, softwares, or locked in to a certain vendor or service provider'? I see that being locked in is mentioned later in the document.
Line 116 - changed "thousands" to plural. 
Line 269 - Is 'shenanigans' too informal for a GC strategy document?